% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_myb_file.R
\name{clean_myb_file}
\alias{clean_myb_file}
\title{Permet de clean un fichier excel du myb de façon standardisée}
\usage{
clean_myb_file(excel_path)
}
\arguments{
\item{excel_path}{Un chemin d'accès vers un fichier excel du myb}
}
\value{
Un dataframe du myb clean
}
\description{
Permet de clean un fichier excel du myb de façon standardisée
}
\examples{
# Pas d'exemple
}
