% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db_myb.R
\name{create_db_myb}
\alias{create_db_myb}
\title{Permet de créer la databse myb sur la production des minerais à partir des données déjà téléchargées.}
\usage{
create_db_myb(
  folder_data_path = here::here("01-data"),
  nb_workers = 4,
  path_db = here::here("03-output", "database-mineral-production.csv")
)
}
\arguments{
\item{folder_data_path}{Chemin d'accès à l'endroit où sont stockées les données.}

\item{nb_workers}{Nombre de travailleurs.}

\item{path_db}{Chemin d'accès pour enregistrer la database.}
}
\value{
Un fichier csv.
}
\description{
Permet de créer la databse myb sur la production des minerais à partir des données déjà téléchargées.
}
\examples{
# Pas d'exemple.
}
