% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_myb.R
\name{db_myb}
\alias{db_myb}
\title{Permet de lancer toute la procédure de téléchargement et de précleaning de la base de données myb.}
\usage{
db_myb(
  path_usgs_myb =
    "https://www.usgs.gov/centers/national-minerals-information-center/minerals-yearbook-metals-and-minerals",
  nb_workers = 4,
  folder_path_data = here::here("01-data"),
  path_db = here::here("03-output", "database-mineral-production.csv")
)
}
\arguments{
\item{path_usgs_myb}{Lien vers la page d'acceuil myb de USGS.}

\item{nb_workers}{Le nombre de travailleurs.}

\item{folder_path_data}{Chemin d'accès au répertoire qui contiendra les données téléchargées.}

\item{path_db}{Chemin d'accès pour l'enregistrement du fichier csv de la base de données.}
}
\value{
Un fichier csv.
}
\description{
Permet de lancer toute la procédure de téléchargement et de précleaning de la base de données myb.
}
\examples{
# Pas d'exemple.
}
