% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_all_myb.R
\name{download_all_myb}
\alias{download_all_myb}
\title{Permet de récupérer les liens de toutes les pages du myb et de lancer les téléchargement des fichiers Excel.}
\usage{
download_all_myb(
  path_usgs_myb =
    "https://www.usgs.gov/centers/national-minerals-information-center/minerals-yearbook-metals-and-minerals",
  nb_workers = 4
)
}
\arguments{
\item{path_usgs_myb}{Un lien vers la page d'acceuil du myb de USGS. Le lien est rentré par défaut.}

\item{nb_workers}{Nombre de travailleurs.}
}
\value{
Tous les fichiers Excel du myb.
}
\description{
Permet de récupérer les liens de toutes les pages du myb et de lancer les téléchargement des fichiers Excel.
}
\examples{
# Pas d'exemple.
}
