% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_one_file.R
\name{download_one_file}
\alias{download_one_file}
\title{Permet de télécharger un ficheir à partir de son url et d'enregistrer un message d'erreur s'il y en a.}
\usage{
download_one_file(excel_file_link, excel_file_path)
}
\arguments{
\item{excel_file_link}{Une URl vers un fichier.}

\item{excel_file_path}{Un chemin d'accès vers l'endroit où le ficheir doit être enregistré.}
}
\value{
Un fichier téléchargé et/ou un document d'erreur.
}
\description{
Permet de télécharger un ficheir à partir de son url et d'enregistrer un message d'erreur s'il y en a.
}
\examples{
# Pas d'exemple
}
