% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_download_myb.R
\name{launch_download_myb}
\alias{launch_download_myb}
\title{Permet de lancer le processus de téléchargement du myb en limitant les risques d'erreur grâce à un fichier compilant les messages d'erreur.}
\usage{
launch_download_myb(
  path_usgs_myb =
    "https://www.usgs.gov/centers/national-minerals-information-center/minerals-yearbook-metals-and-minerals",
  nb_workers = 4
)
}
\arguments{
\item{path_usgs_myb}{Un lien vers la page d'acceuil du myb de USGS. Le lien est rentré par défaut.}

\item{nb_workers}{Le nombre de travailleurs.}
}
\value{
Tous les fichiers Excel du myb.
}
\description{
Permet de lancer le processus de téléchargement du myb en limitant les risques d'erreur grâce à un fichier compilant les messages d'erreur.
}
\examples{
# Pas d'exemple.
}
