% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_empty_file.R
\name{rm_empty_file}
\alias{rm_empty_file}
\title{Permet de supprimer les fichiers vides inutiles.}
\usage{
rm_empty_file(excel_path)
}
\arguments{
\item{excel_path}{Un chemin d'accès vers un fichier.}
}
\value{
Des dossiers supprimés.
}
\description{
Permet de supprimer les fichiers vides inutiles.
}
\examples{
# Pas d'exemple
}
