% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_one_file.R
\name{download_one_file}
\alias{download_one_file}
\title{Allow to download a file with is URL if it doesn't exist already. Write an error message in a document if there is an error.}
\usage{
download_one_file(excel_file_link, excel_file_path)
}
\arguments{
\item{excel_file_link}{An url link leading to an excel file}

\item{excel_file_path}{A path indicating where the file should be save}
}
\value{
A saving file or an error document
}
\description{
Allow to download a file with is URL if it doesn't exist already. Write an error message in a document if there is an error.
}
\examples{
#No example
}
